requirements.txt
simpletransformers~=0.63.6
torch
detoxify
ftfy==6.1.1
praw==7.6.0
sacremoses
azure-functions
azure-storage-queue
azure-data-tables
dataclasses~=0.8
asyncpraw~=7.5.0
aiohttp~=3.8.1
sqlalchemy~=1.4.37